% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cip_srs.R
\name{cip_srs}
\alias{cip_srs}
\title{Confidence interval of the proportion estimated from a Simple Random Sampling plan}
\usage{
cip_srs(C, e, p_exp = 0.5, n_real, p_est, N = Inf)
}
\arguments{
\item{C}{Level of confidence. (0 <= C <= 1)}

\item{e}{Sampling error. (0 <= e <= 1).}

\item{p_exp}{Expected proportion in the population. By default 0.5. (0 <= p <= 1)}

\item{n_real}{Real sample size. (n > 0).}

\item{p_est}{Sample proportion. (0 <= p <= 1)}

\item{N}{Population size, by default is infinite. Must be a positive integer.}
}
\value{
The function returns the sample size needed to estimate the proportion of occurrence of a phenomena, consistent with the risk ('C' and 'e') that the auditor is willing to assume, when conducting a Simple Random Sampling plan, among others with some restrictions.
}
\description{
Confidence interval of the proportion estimated from a Simple Random Sampling plan
}
\examples{
np_srs(0.90, 0.04)
np_srs(0.90, 0.04, 0.5)
np_srs(0.90, 0.04, 0.5, 20000)
np_srs(0.90, 0.04, N = 20000)
}
