% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_cls.R
\name{cp_cls}
\alias{cp_cls}
\title{Confidence interval of the proportion under a cluster sampling design}
\usage{
cp_cls(C, p_est, n_real, sd_est, m, N = Inf, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{p_est}{Proportion mean.}

\item{n_real}{Real sample size (id est, the number of clusters sampled); n_real > 0.}

\item{sd_est}{Estimated standard deviation; sd_est > 0.}

\item{m}{Average cluster size (number of elements), either of the population -preferred- or of a preliminary sample; m > 0.}

\item{N}{A positive integer indicating the population size (id est, the total number of clusters). Defaults to infinite.}

\item{parameter}{Type TRUE if you do know the populations sd, type FALSE (default) if it is an estimate.}
}
\value{
This function returns the confidence interval of the population proportion when using a cluster sampling design without replacement, given the sample size.
}
\description{
Confidence interval of the proportion under a cluster sampling design
}
\details{
The function to calculate the confidence interval is:
\deqn{CI = \hat{p} \pm MOE}
where:
\deqn{MOE = Z \cdot \sqrt{ \frac{p.(1 - p)}{(n - 1)} \cdot \frac{(N - n)}{N \cdot m^2} }}
\eqn{\hat{p}} is the sample proportion, 'sd' is parameter 'sd_est', and 'Z' is the quantile of the two-tailed normal distribution function, compatible with the chosen confidence level 'C'.
If 'sd_est' is unknown, the t-student is used instead of the normal distribution.
}
\examples{
cp_cls(C = 0.95, p_est = 0.48, n_real = 25, N = 415, m = 8, sd_est = 0.73, parameter = TRUE)
cp_cls(C = 0.95, p_est = 0.48, n_real = 19, N = 500, m = 8, sd_est = 0.33, parameter = TRUE)
}
