% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_sts.R
\name{cp_sts}
\alias{cp_sts}
\title{Confidence interval of the proportion under a Stratified Sampling framework}
\usage{
cp_sts(C, n, p_est, N)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{n}{A vector with the real sample size for each stratum (n > 0)}

\item{p_est}{A vector with the sample proportion of each stratum (0 <= p_est <= 1)}

\item{N}{A vector indicating the number of elements in each stratum (N > 0).}
}
\value{
The function returns the interval of confidence of the population proportion.
}
\description{
Confidence interval of the proportion under a Stratified Sampling framework
}
\examples{
cp_sts(C = 0.95, p_est = c(0.3, 0.5, 0.7), n = c(100, 150, 200), N = c(200, 250, 300))
}
