% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_cls.R
\name{ct_cls}
\alias{ct_cls}
\title{Confidence interval of the total under a cluster sampling design}
\usage{
ct_cls(C, x_est, n_real, sd_est, m, N = Inf, M = NULL, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{x_est}{Sample mean.}

\item{n_real}{Real sample size (id est, the number of clusters sampled); n_real > 0.}

\item{sd_est}{Estimated standard deviation; sd_est > 0.}

\item{m}{Average cluster size (number of elements), either of the population -preferred- or of a preliminary sample; m > 0.}

\item{N}{A positive integer indicating the population size (id est, the total number of clusters). Defaults to infinite.}

\item{M}{Number of elements in the population; M > 0. Leave it Null if unknown.}

\item{parameter}{Type TRUE if you do know the populations sd, type FALSE (default) if it is an estimate.}
}
\value{
This function returns the confidence interval of the population total when using a cluster sampling design without replacement, given the sample size.
}
\description{
Confidence interval of the total under a cluster sampling design
}
\examples{
ct_cls(C = 0.95, x_est = 9990, n_real = 30, N = 500, m = 8, M = 2500, sd_est = 15000, TRUE)
ct_cls(C = 0.95, x_est = 9990, n_real = 19, N = 500, m = 8, sd_est = 15000, parameter = TRUE)
}
