% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_sts.R
\name{ct_sts}
\alias{ct_sts}
\title{Confidence interval of the total under a Stratified Sampling framework}
\usage{
ct_sts(C, n, x_est, sd_est, parameter = FALSE, N)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{n}{A vector with the real sample size for each stratum (n > 0)}

\item{N}{A vector with the number of elements in each stratum (N > 0).}

\item{t_est}{A vector with the sample total of each stratum (0 <= p_est <= 1)}
}
\value{
The function returns the interval of confidence of the population total
}
\description{
Confidence interval of the total under a Stratified Sampling framework
}
\examples{
ct_sts(C = 0.95, x_est = c(0.3, 0.5, 0.7), n = c(100, 150, 200), N = c(200, 250, 300))
}
