% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_sts.R
\name{cx_sts}
\alias{cx_sts}
\title{Confidence interval of the mean under a Stratified Sampling framework}
\usage{
cx_sts(C, n, x_est, sd_est, parameter = FALSE, N)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{n}{A vector with the real sample size for each stratum (n > 0)}

\item{x_est}{A vector with the sample mean of each stratum (0 <= p_est <= 1)}

\item{N}{A vector with the number of elements in each stratum (N > 0).}
}
\value{
The function returns the interval of confidence of the population mean.
}
\description{
Confidence interval of the mean under a Stratified Sampling framework
}
\examples{
cx_sts(C = 0.95, n = c(100, 150, 200), x_est = c(3.2, 3.5, 3.7), sd_est = c(0.5, 0.6, 0.4), parameter = TRUE, N = c(200, 250, 300))
}
