% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_cls.R
\name{ep_cls}
\alias{ep_cls}
\title{Sampling error when estimating the proportion using a cluster sampling design}
\usage{
ep_cls(C, n_real, sd_est, m, N = Inf, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{n_real}{A positive integer representing the real sample size (the number of clusters sampled); n_real > 0.}

\item{sd_est}{Estimated standard deviation ; sd_est > 0.}

\item{m}{Average cluster size (number of elements), either of the population -preferred- or of a preliminary sample; m > 0.}

\item{N}{A positive integer indicating the population size (id est, the total number of clusters). Defaults to infinite.}

\item{parameter}{Type TRUE if you do know the populations SD in sd_exp, or type FALSE (default) if they are estimates.}
}
\value{
This function returns the sampling error when using a cluster sampling design without replacement to estimate the proportion, given the sample size.
}
\description{
Sampling error when estimating the proportion using a cluster sampling design
}
\examples{
ep_cls(C = 0.95, n_real = 31, sd_est = 0.726, m = 6.04, N = 415, parameter = TRUE)
}
