% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_srs.R
\name{ep_srs}
\alias{ep_srs}
\title{Sampling error when estimating the proportion with a Simple Random Sampling plan}
\usage{
ep_srs(C, n, p = 0.5, N = Inf)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{n}{Given sample size}

\item{p}{Expected or estimated (preferred) proportion in the population.}

\item{N}{Population size. Must be a positive integer.}
}
\value{
The function returns the sample error consistent with the estimation of the proportion of occurrence of a phenomena, given the sample size.
}
\description{
Sampling error when estimating the proportion with a Simple Random Sampling plan
}
\examples{
ep_srs(0.90, 415)
ep_srs(0.90, 415, 0.5)
ep_srs(0.90, 415, 0.5, 20000)
ep_srs(0.90, 415, N = 20000)
}
