% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_cls.R
\name{et_cls}
\alias{et_cls}
\title{Sampling error when estimating the total using a cluster sampling design}
\usage{
et_cls(C, n_real, sd_est, N = Inf, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{n_real}{A positive integer representing the real sample size (the number of clusters sampled); n_real > 0.}

\item{sd_est}{Estimated standard deviation ; sd_est > 0.}

\item{N}{A positive integer indicating the population size (id est, the total number of clusters). Defaults to infinite.}

\item{parameter}{Type TRUE if you do know the populations SD in sd_exp, or type FALSE (default) if they are estimates.}
}
\value{
This function returns the sampling error when using a cluster sampling design without replacement to estimate the total, given the sample size.
}
\description{
Sampling error when estimating the total using a cluster sampling design
}
\details{
The function looks for the value of 'E' that, given the real sampling size, fits:
\deqn{n = \frac{N \cdot \text{sd}^2}{\frac{E^2}{N \cdot Z^2} + \text{sd}^2}}
where 'sd' is parameter 'sd_est', and 'Z' is the quantile of the two-tailed normal distribution function, compatible with the chosen confidence level 'C'.
If 'sd_exp' is unknown, the t-student is used instead of the normal distribution.
}
\examples{
et_cls(C = 0.95, n_real = 213, sd_est = 25189, N = 415, parameter = TRUE)
}
