% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_srs.R
\name{et_srs}
\alias{et_srs}
\title{Sampling error when estimating the total with a Simple Random Sampling plan}
\usage{
et_srs(C, n_real, sd_exp, parameter = FALSE, N)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{n_real}{Given sample size (n_real > 0)}

\item{sd_exp}{Expected standard deviation (sd_exp > 0)}

\item{parameter}{Type TRUE if you do know the populations sd, type FALSE (default) if it is an estimate.}

\item{N}{A positive integer indicating the number of elements in the population.}
}
\value{
The function returns the sample error consistent with the estimation of the total, given the sample size.
}
\description{
Sampling error when estimating the total with a Simple Random Sampling plan
}
\examples{
ex_srs(C = 0.95, n_real = 87, sd_exp = 4.1, N = 1200)
ex_srs(C = 0.95, n_real = 23, sd_exp = 4.1, N = 1200)
ex_srs(C = 0.95, n_real = 6, sd_exp = 2.1, parameter = TRUE, N = 1200)
}
