% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_sts.R
\name{ex_sts}
\alias{ex_sts}
\title{Sampling error when estimating the mean with a Stratified Sampling plan}
\usage{
ex_sts(C, n_real, sd_est, alloc = NULL, parameter = FALSE, N)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{n_real}{Given sample size (n_real > 0)}

\item{parameter}{Type TRUE if you do know the populations SD, type FALSE (default) if it is an estimate.}

\item{N}{A positive integer indicating the number of elements in each strata. # ###REVISAR}

\item{p_est}{Sample proportion (0 <= p_est <= 1).}
}
\value{
The function returns the sample error consistent with the estimation of the mean of a phenomena, given the sample size.
}
\description{
Sampling error when estimating the mean with a Stratified Sampling plan
}
\examples{
ex_sts(C = 0.95, n_real = c(48, 14, 7), sd_est = c(0.2, 0.5, 0.7), alloc =c(0.7, 0.2, 0.1), parameter = TRUE, N = c(1400, 400, 200))
}
