% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_srs.R
\name{nm_srs}
\alias{nm_srs}
\title{Sample size 'n' needed to estimate the mean 'm' with a Simple Random Sampling (SRS) plan, given the sampling error 'E'}
\usage{
nm_srs(C, E, s, N = Inf)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{E}{Sampling error (E > 1).}

\item{s}{Standard deviation}

\item{N}{Population size, by default is infinite. Must be a positive integer.}
}
\value{
The function returns the sample size needed to estimate the mean of a phenomena, consistent with the risk ('C' and 'e') that the auditor is willing to assume.
}
\description{
Sample size 'n' needed to estimate the mean 'm' with a Simple Random Sampling (SRS) plan, given the sampling error 'E'
}
\examples{
nm_srs(0.95, 50, 400)
nm_srs(0.95, 50, 400, 10000)
nm_srs(0.95, 100, 400, 10000)
}
