% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_srs.R
\name{np_srs}
\alias{np_srs}
\title{Sample size 'n' needed to estimate the proportion 'p' with a Simple Random Sampling (SRS) plan, given the sampling error 'e'}
\usage{
np_srs(C, e, p = 0.5, N = Inf)
}
\arguments{
\item{C}{Level of confidence. (0 <= C <= 1)}

\item{e}{Sampling error. (0 <= e <= 1).}

\item{p}{Expected proportion in the population. By default 0.5, also applicable if you lack information. (0 <= p <= 1)}

\item{N}{Population size, by default is infinite. Must be a positive integer.}
}
\value{
The function returns the sample size needed to estimate the proportion of occurrence of a phenomena, consistent with the risk ('C' and 'e') that the auditor is willing to assume, when conducting a Simple Random Sampling plan, among others with some restrictions.
}
\description{
Sample size 'n' needed to estimate the proportion 'p' with a Simple Random Sampling (SRS) plan, given the sampling error 'e'
}
\examples{
np_srs(0.90, 0.04)
np_srs(0.90, 0.04, 0.5)
np_srs(0.90, 0.04, 0.5, 20000)
np_srs(0.90, 0.04, N = 20000)
}
