% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_sts.R
\name{np_sts}
\alias{np_sts}
\title{Sample size required to estimate the proportion using a stratified sampling design}
\usage{
np_sts(C, e, p_exp, alloc = NULL, N)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{e}{Sampling error; 0 <= e <= 1.}

\item{p_exp}{A vector with the expected proportion in each stratum; 0 <= p_exp(i) <= 1, for every 'i'.}

\item{alloc}{A vector with the relative allocation of sample size for each stratum; 0 < alloc(i) < 1, where sum(alloc(i)) = 1. If not defined (default), its values would be proportional to the size of each stratum.}

\item{N}{A vector of positive integers representing the number of elements in each stratum.}
}
\value{
This function returns the sample size required to estimate the proportion of occurrences of an event when using a stratified sampling design without replacement, given the level of risk.
}
\description{
Sample size required to estimate the proportion using a stratified sampling design
}
\examples{
np_sts(C = 0.95, e = 0.1, p_exp = c(0.2, 0.3), N = c(220, 350))
np_sts(C = 0.95, e = 0.1, p_exp = c(0.1, 0.5, 0.5), N = c(150, 40, 110))
}
