% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nt_sts.R
\name{nt_sts}
\alias{nt_sts}
\title{Sample size required to estimate the total using a stratified sampling design}
\usage{
nt_sts(C, E, sd_exp, alloc = NULL, N, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{E}{Sampling error; E > 0.}

\item{sd_exp}{A vector with the expected standard deviation in each stratum; sd_exp(i) > 0, for every 'i' stratum.}

\item{alloc}{A vector with the relative allocation of sample size for each stratum; 0 < alloc(i) < 1, where sum(alloc(i)) = 1. If not defined (default), its values would be proportional to the size of each stratum.}

\item{N}{A vector of positive integers representing the number of elements in each stratum.}

\item{parameter}{Type TRUE if you do know the populations SD in sd_exp, or type FALSE (default) if they are estimates.}
}
\value{
This function returns the sample size required to estimate the mean of a variable when using a stratified sampling design without replacement, given the level of risk.
}
\description{
Sample size required to estimate the total using a stratified sampling design
}
\details{
The function to calculate the sample size is:
\deqn{n = \frac{\sum_{i=1}^{s} \frac{N_i^2 \cdot \text{sd}_i^2}{\text{alloc}_i} }{ N^2 \cdot \frac{E^2}{N^2 \cdot Z^2} + \sum_{i=1}^{s} N_i \cdot \text{sd}_i^2 }}
where 'sd' is parameter 'sd_exp', and 'Z' is the quantile of the two-tailed normal distribution function, compatible with the chosen confidence level 'C'.
If 'sd_exp' is unknown, the t-student is used instead of the normal distribution.
}
\examples{
nt_sts(C = 0.95, E = 200, sd_exp = c(5, 15), alloc = c(1/2, 1/2), N = c(155, 62), parameter = TRUE)
nt_sts(C = 0.95, E = 400, sd_exp = c(5, 15, 10), alloc = c(1/2, 1/4, 1/4), N = c(150, 40, 110), parameter = FALSE)
}
