% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nx_srs.R
\name{nx_srs}
\alias{nx_srs}
\title{Sample size required to estimate the mean using a simple random sampling design}
\usage{
nx_srs(C, E, sd_exp, N = Inf, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{E}{Sampling error; E > 0.}

\item{sd_exp}{Expected standard deviation; sd_exp > 0.}

\item{N}{A positive integer representing the number of elements in the population. Defaults to infinite.}

\item{parameter}{Type TRUE if you do know the population SD, or type FALSE (default) if it is an estimate.}
}
\value{
This function returns the sample size required to estimate the mean of a variable when using a simple random sampling design without replacement, given the level of risk.
}
\description{
Sample size required to estimate the mean using a simple random sampling design
}
\details{
The function to calculate the sample size is:
\deqn{n = \frac{N \cdot \text{sd}^2}{(N - 1) \cdot \frac{E^2}{Z^2} + \text{sd}^2 }}
where 'sd' is parameter 'sd_exp', and Z' is the quantile of the two-tailed normal distribution function, compatible with the chosen confidence level 'C'.
If 'sd_exp' is unknown, the t-student is used instead of the normal distribution.
}
\examples{
nx_srs(C = 0.95, E = 50, sd_exp = 400)
nx_srs(C = 0.95, E = 50, sd_exp = 400, parameter = TRUE, N = 10000)
nx_srs(C = 0.95, E = 50, sd_exp = 400, N = 10000, parameter = FALSE)
nx_srs(C = 0.95, E = 100, sd_exp = 400, N = 10000)


}
