% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paskal-package.R
\docType{package}
\name{Paskal-package}
\alias{Paskal}
\alias{Paskal-package}
\title{Paskal: Performance audit sample kit, along other useful things}
\description{
The Paskal package assists performance auditors with useful statistical sampling calculations. It includes functions to calculate sample size, sampling error, and confidence intervals when estimating proportions, means, or totals through various sampling plans.
}
\keyword{internal}
