% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample4audit-package.R
\docType{package}
\name{sample4audit-package}
\alias{sample4audit}
\alias{sample4audit-package}
\title{sample4audit: Performance audit sample kit, along lots of other useful things}
\description{
The Paskal package was designed to assist performance auditors with some useful calculations related to statistical sampling, including functions to calculate the sample size, the sampling error, or the confidence interval; when estimating the proportion, the mean, or the total thorigh differente sampling plans.
}
\keyword{internal}
