% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_sts.R
\name{cp_sts}
\alias{cp_sts}
\title{Confidence interval of the proportion under a stratified sampling design}
\usage{
cp_sts(C, n_real, p_est, N)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{n_real}{A vector of positive integers representing the real sample size of each stratum; n_real(i) > 0, for every 'i' stratum.}

\item{p_est}{A vector with the estimated proportion in each stratum; 0 <= p_exp(i) <= 1.}

\item{N}{A vector of positive integers representing the number of elements in each stratum; N(i) > 0.}
}
\value{
This function returns the global confidence interval when using a stratified sampling design without replacement to estimate the proportion, given the sample size.
}
\description{
Confidence interval of the proportion under a stratified sampling design
}
\details{
The function to calculate the confidence interval is:
\deqn{CI = \hat{p} \pm MOE}
where:
\deqn{MOE = Z \cdot \sqrt{ \frac{1}{N^2} \cdot \sum_{i=1}^{s} N_i^2  \cdot \frac{p_i.(1 - p_i)}{(n_i - 1)} \cdot \frac{(N_i - n_i)}{N_i} }}
\eqn{\hat{p}} is the sample proportion and 'Z' is the quantile of the two-tailed normal distribution function, compatible with the chosen confidence level 'C'.
}
\examples{
cp_sts(C = 0.95, n_real = c(100, 150, 200), p_est = c(0.3, 0.5, 0.7), N = c(200, 250, 300))
}
