% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_srs.R
\name{ct_srs}
\alias{ct_srs}
\title{Confidence interval of the total under a simple random sampling design}
\usage{
ct_srs(C, sd_est, t_est, n_real, parameter = FALSE, N = Inf)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{sd_est}{Expected standard deviation; s_exp > 0.}

\item{t_est}{Sample total.}

\item{n_real}{Real sample size; n_real > 0.}

\item{parameter}{Type TRUE if you do know the populations sd, type FALSE (default) if it is an estimate.}

\item{N}{A positive integer indicating the number of elements in the population. Defaults to infinite.}
}
\value{
The function returns the confidence interval of the population total.
}
\description{
Confidence interval of the total under a simple random sampling design
}
\details{
The function to calculate the confidence interval is:
\deqn{CI = \hat{t} \pm MOE}
where:
\deqn{MOE = Z \cdot \sqrt{ \frac{\text{sd}^2}{n} \cdot \frac{(N - n)} {N} \cdot N^2}}
\eqn{\hat{t}} is the sample total, 'sd' is parameter 'sd_est', and 'Z' is the quantile of the two-tailed normal distribution function, compatible with the chosen confidence level 'C'.
If 'sd_est' is unknown, the t-student is used instead of the normal distribution.
}
\examples{
ct_srs(C = 0.95, t_est = 3500, sd_est = 4.1, n_real = 87, N = 1200)
ct_srs(C = 0.95, t_est = 3500, sd_est = 4.1, n_real = 87, N = 1200, parameter = TRUE)
}
