% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_sts.R
\name{ct_sts}
\alias{ct_sts}
\title{Confidence interval of the total under a stratified sampling design}
\usage{
ct_sts(C, n_real, x_est, sd_est, N, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{n_real}{A vector of positive integers representing the real sample size of each stratum; n_real(i) > 0, for every 'i' stratum.}

\item{x_est}{A vector with the sample mean of each stratum.}

\item{sd_est}{A vector with the expected standard deviation in each stratum; sd_exp(i) > 0, for every 'i' stratum.}

\item{N}{A vector of positive integers representing the number of elements in each stratum; N(i) > 0.}

\item{parameter}{Type TRUE if you do know the populations SD in sd_exp, or type FALSE (default) if they are estimates.}
}
\value{
This function returns the global confidence interval when using a stratified sampling design without replacement to estimate the total, given the sample size.
}
\description{
Confidence interval of the total under a stratified sampling design
}
\details{
The function to calculate the confidence interval is:
\deqn{CI = \hat{t} \pm MOE}
where:
\deqn{MOE = Z \cdot \sqrt{ \sum_{i=1}^{s} N_i^2  \cdot \frac{\text{sd}_i^2}{n_i} \cdot \frac{(N_i - n_i)}{N_i}  }}
\eqn{\hat{t}} is the sample total, 'sd' is parameter 'sd_est', and 'Z' is the quantile of the two-tailed normal distribution function, compatible with the chosen confidence level 'C'.
If 'sd_est' is unknown, the t-student is used instead of the normal distribution.
}
\examples{
ct_sts(C = 0.95, n_real = c(100, 150, 200), x_est = c(0.3, 0.5, 0.7), sd_est = c(0.5, 0.6, 0.4), N = c(200, 250, 300))
}
