% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_cls.R
\name{cx_cls}
\alias{cx_cls}
\title{Confidence interval of the mean under a cluster sampling design}
\usage{
cx_cls(C, x_est, n_real, sd_est, m, N = Inf, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{x_est}{Sample mean.}

\item{n_real}{Real sample size (id est, the number of clusters sampled); n_real > 0.}

\item{sd_est}{Estimated standard deviation; sd_est > 0.}

\item{m}{Average cluster size (number of elements), either of the population -preferred- or of a preliminary sample; m > 0.}

\item{N}{A positive integer indicating the population size (id est, the total number of clusters). Defaults to infinite.}

\item{parameter}{Type TRUE if you do know the populations sd, type FALSE (default) if it is an estimate.}
}
\value{
This function returns the confidence interval of the population mean when using a cluster sampling design without replacement, given the sample size.
}
\description{
Confidence interval of the mean under a cluster sampling design
}
\details{
The function to calculate the confidence interval is:
\deqn{CI = \hat{x} \pm MOE}
where:
\deqn{MOE = Z \cdot \sqrt{ \frac{\text{sd}^2}{n} \cdot \frac{(N - n)}{N \cdot m^2} }},
\eqn{\hat{x}} is the sample mean, 'sd' is parameter 'sd_est', and 'Z' is the quantile of the two-tailed normal distribution function, compatible with the chosen confidence level 'C'.
If 'sd_est' is unknown, the t-student is used instead of the normal distribution.
}
\examples{
cx_cls(C = 0.95, x_est = 9990, n_real = 30, N = 500, m = 8, sd_est = 15000, parameter = TRUE)
cx_cls(C = 0.95, x_est = 9990, n_real = 19, N = 500, m = 8, sd_est = 15000, parameter = TRUE)
}
