% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_srs.R
\name{cx_srs}
\alias{cx_srs}
\title{Confidence interval of the mean under a simple random sampling design}
\usage{
cx_srs(C, sd_est, x_est, n_real, parameter = FALSE, N = Inf)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{sd_est}{Estimated standard deviation; s_exp > 0.}

\item{x_est}{Sample mean.}

\item{n_real}{Real sample size; n_real > 0.}

\item{parameter}{Type TRUE if you do know the populations sd, type FALSE (default) if it is an estimate.}

\item{N}{A positive integer indicating the number of elements in the population. Defaults to infinite.}
}
\value{
The function returns the confidence interval of the population mean.
}
\description{
Confidence interval of the mean under a simple random sampling design
}
\details{
The function to calculate the confidence interval is:
\deqn{CI = \hat{x} \pm MOE}
where:
\deqn{MOE = Z \cdot \sqrt{\frac{\text{sd}^2}{n} \cdot \frac{(N - n)} {N}}},
\eqn{\hat{x}} is the sample mean, 'sd' is parameter 'sd_est', and 'Z' is the quantile of the two-tailed normal distribution function, compatible with the chosen confidence level 'C'.
If 'sd_est' is unknown, the t-student is used instead of the normal distribution.
}
\examples{
cx_srs(C = 0.95, sd_est = 200, x_est = 3500, n_real = 250)
cx_srs(C = 0.95, sd_est = 200, x_est = 3500, n_real = 400)
cx_srs(C = 0.95, sd_est = 200, x_est = 3500, n_real = 250, N = 5000, parameter = TRUE)
}
