% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_srs.R
\name{em_srs}
\alias{em_srs}
\title{Sampling error when estimating the mean with a Simple Random Sampling plan}
\usage{
em_srs(C, n, s, N = Inf)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{n}{Given sample size}

\item{s}{Standard deviation}

\item{N}{Population size. Must be a positive integer.}
}
\value{
The function returns the sample error consistent with the estimation of the mean, given the sample size.
}
\description{
Sampling error when estimating the mean with a Simple Random Sampling plan
}
\examples{
em_srs(0.95, 140, 600)
em_srs(0.95, 140, 800)
em_srs(0.95, 140, 800, 2000)
}
