% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_srs.R
\name{ep_srs}
\alias{ep_srs}
\title{Sample error e for simple random sampling of proportion}
\usage{
ep_srs(C, n, p, N = Inf)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{n}{Given sample size}

\item{p}{Expected or estimated (preferred) proportion in the population.}

\item{N}{Population size. Must be a positive integer.}
}
\value{
The function returns the sample error consistent with the estimation of the proportion of occurrende of a phenomena, given the sample size.
}
\description{
Sample error e for simple random sampling of proportion
}
\examples{
ep_srs(0.95, 151, 0.5, 9999999999)
}
