% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_srs.R
\name{ep_srs}
\alias{ep_srs}
\title{Sampling error when estimating the proportion using a simple random sampling design}
\usage{
ep_srs(C, n_real, p_est = 0.5, N = Inf)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{n_real}{Real sample size; n_real > 0.}

\item{p_est}{Sample proportion; 0 <= p_est <= 1.}

\item{N}{A positive integer representing the number of elements in the population. Defaults to infinite.}
}
\value{
This function returns the sampling error when using a simple random sampling design without replacement to estimate the proportion, given the sample size.
}
\description{
Sampling error when estimating the proportion using a simple random sampling design
}
\examples{
ep_srs(C = 0.90, n_real = 415)
ep_srs(C = 0.90, n_real = 415, p_est = 0.4)
ep_srs(C = 0.90, n_real = 415, p_est = 0.4, N = 10000)
ep_srs(C = 0.90, n_real = 415, N = 10000)
}
