% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_sts.R
\name{ep_sts}
\alias{ep_sts}
\title{Sampling error when estimating the proportion with a Stratified Sampling plan}
\usage{
ep_sts(C, n_real, p_est, alloc = NULL, parameter = FALSE, N)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{n_real}{Given sample size (n_real > 0)}

\item{p_est}{Sample proportion (0 <= p_est <= 1).}

\item{parameter}{Type TRUE if you do know the populations SD, type FALSE (default) if it is an estimate. # ###REVISAR}

\item{N}{A positive integer indicating the number of elements in each strata. # ###REVISAR}
}
\value{
The function returns the sample error consistent with the estimation of the proportion of occurrence of a phenomena, given the sample size.
}
\description{
Sampling error when estimating the proportion with a Stratified Sampling plan
}
\examples{
ep_sts(C = 0.95, n_real = c(66, 9, 9), p_est = c(0.2, 0.5, 0.7), alloc = c(0.8, 0.1, 01), parameter = TRUE, N = c(1400, 400, 200))
}
