% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_sts.R
\name{et_sts}
\alias{et_sts}
\title{Sampling error when estimating the total with a Stratified Sampling plan}
\usage{
et_sts(C, n_real, sd_est, alloc = NULL, parameter = FALSE, N)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{n_real}{Given sample size (n_real > 0)}

\item{parameter}{Type TRUE if you do know the populations SD, type FALSE (default) if it is an estimate.}

\item{N}{A positive integer indicating the number of elements in each strata. # ###REVISAR}

\item{p_est}{Sample proportion (0 <= p_est <= 1).}
}
\value{
The function returns the sample error consistent with the estimation of the total of a phenomena, given the sample size.
}
\description{
Sampling error when estimating the total with a Stratified Sampling plan
}
\examples{
et_sts(C = 0.95, n_real = c(48, 14, 7), sd_est = c(5, 15, 10), alloc = c(1/2, 1/4, 1/4), parameter = TRUE, N = c(155, 62, 93))
}
