% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_sts.R
\name{et_sts}
\alias{et_sts}
\title{Sampling error when estimating the total using a stratified sampling design}
\usage{
et_sts(C, n_real, sd_est, alloc = NULL, N, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{n_real}{A vector of positive integers representing the real sample size of each stratum.}

\item{sd_est}{A vector with the estimated standard deviation in each stratum; sd_est(i) > 0, for every 'i' stratum.}

\item{alloc}{A vector with the relative allocation of sample size for each stratum; 0 < alloc(i) < 1, where sum(alloc(i)) = 1. If not defined (default), its values would be proportional to the size of each stratum.}

\item{N}{A vector of positive integers representing the number of elements in each stratum; N(i) > 0.}

\item{parameter}{Type TRUE if you do know the populations SD in sd_exp, or type FALSE (default) if they are estimates.}
}
\value{
This function returns the global sampling error when using a stratified sampling design without replacement to estimate the total, given the sample size.
}
\description{
Sampling error when estimating the total using a stratified sampling design
}
\details{
The function looks for the value of 'E' that, given the real sampling size for each stratum, fits:
\deqn{n = \frac{\sum_{i=1}^{s} \frac{N_i^2 \cdot \text{sd}_i^2}{\text{alloc}_i} }{ N^2 \cdot \frac{E^2}{N^2 \cdot Z^2} + \sum_{i=1}^{s} N_i \cdot \text{sd}_i^2 }}
where 'sd' is parameter 'sd_est', and 'Z' is the quantile of the two-tailed normal distribution function, compatible with the chosen confidence level 'C'.
If 'sd_exp' is unknown, the t-student is used instead of the normal distribution.
}
\examples{
et_sts(C = 0.95, n_real = c(57, 29, 29), sd_est = c(5, 15, 10), alloc = c(1/2, 1/4, 1/4), N = c(150, 40, 110), parameter = FALSE)
}
