% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_srs.R, R/ex_srs.R
\name{ex_srs}
\alias{ex_srs}
\title{Sampling error when estimating the total with a Simple Random Sampling plan}
\usage{
ex_srs(C, n_real, sd_exp, N = Inf)

ex_srs(C, n_real, sd_exp, N = Inf)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{n_real}{Given sample size (n_real > 0)}

\item{sd_exp}{Expected standard deviation (sd_exp > 0)}

\item{N}{A positive integer indicating the number of elements in the population. By default, infinite.}
}
\value{
The function returns the sample error consistent with the estimation of the mean, given the sample size.

The function returns the sample error consistent with the estimation of the mean, given the sample size.
}
\description{
Sampling error when estimating the total with a Simple Random Sampling plan

Sampling error when estimating the mean with a Simple Random Sampling plan
}
\examples{
ex_srs(C = 0.95, n_real = 87, sd_exp = 4.1, N = 1200)
ex_srs(C = 0.95, n_real = 23, sd_exp = 4.1, N = 1200)
ex_srs(C = 0.95, n_real = 6, sd_exp = 2.1, N = 1200)
ex_srs(C = 0.95, n_real = 140, sd_exp = 600)
ex_srs(C = 0.95, n_real = 140, sd_exp = 800)
ex_srs(C = 0.95, n_real = 140, sd_exp = 800, N = 2000)
}
