% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_sts.R
\name{ex_sts}
\alias{ex_sts}
\title{Sampling error when estimating the mean using a stratified sampling design}
\usage{
ex_sts(C, n_real, sd_est, alloc = NULL, N, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{n_real}{A vector of positive integers representing the real sample size of each stratum.}

\item{sd_est}{A vector with the estimated standard deviation in each stratum; sd_est(i) > 0, for every 'i' stratum.}

\item{alloc}{A vector with the relative allocation of sample size for each stratum; 0 < alloc(i) < 1, where sum(alloc(i)) = 1. If not defined (default), its values would be proportional to the size of each stratum.}

\item{N}{A vector of positive integers representing the number of elements in each stratum; N(i) > 0.}

\item{parameter}{Type TRUE if you do know the populations SD in sd_exp, or type FALSE (default) if they are estimates.}
}
\value{
This function returns the global sampling error when using a stratified sampling design without replacement to estimate the mean, given the sample size.
}
\description{
Sampling error when estimating the mean using a stratified sampling design
}
\examples{
ex_sts(C = 0.95, n_real = c(48, 14, 7), sd_est = c(0.2, 0.5, 0.7), alloc =c(0.7, 0.2, 0.1), N = c(1400, 400, 200), parameter = TRUE)
}
