% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_srs.R
\name{nm_srs}
\alias{nm_srs}
\title{Sample size n for simple random sampling of mean}
\usage{
nm_srs(C, e, s, N = Inf)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{e}{Sampling error (0 <= e <= 1).}

\item{s}{To be completed}

\item{N}{Population size, by default is infinite. Must be a positive integer.}
}
\value{
The function returns the sample size needed to estimate the mean of a phenomena, consistent with the risk ('C' and 'e') that the auditor is willing to assume.
}
\description{
Sample size n for simple random sampling of mean
}
\examples{
np_srs(0.95, 0.03)
np_srs(0.95, 0.03, 0.5)
np_srs(0.95, 0.03, 0.5, 1000)
np_srs(0.95, 0.03, N = 1000)
}
