% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_srs.R
\name{np_srs}
\alias{np_srs}
\title{Sample size required to estimate the proportion using a simple random sampling design}
\usage{
np_srs(C, e, p_exp = 0.5, N = Inf)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{e}{Sampling error; 0 <= e <= 1.}

\item{p_exp}{Expected proportion in the population; 0 <= p_exp <= 1. Enter 0.5 (default) if you lack information.}

\item{N}{A positive integer representing the number of elements in the population. Defaults to infinite.}
}
\value{
This function returns the sample size required to estimate the proportion of occurrences of an event when using a simple random sampling design without replacement, given the level of risk.
}
\description{
Sample size required to estimate the proportion using a simple random sampling design
}
\examples{
np_srs(C = 0.90, e = 0.04)
np_srs(C = 0.90, e = 0.1, p_exp = 0.5)
np_srs(C = 0.90, e = 0.1, p_exp = 0.5, N = 1000)
np_srs(C = 0.90, e = 0.1, p_exp = 0.2, N = 1000)
}
