% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_srs.R
\name{np_srs}
\alias{np_srs}
\title{Sample size needed to estimate the proportion with a Simple Random Sampling plan}
\usage{
np_srs(C, e, p_exp = 0.5, parameter = FALSE, N = Inf)
}
\arguments{
\item{C}{Level of confidence. (0 <= C <= 1)}

\item{e}{Sampling error. (0 <= e <= 1).}

\item{p_exp}{Expected proportion in the population. By default 0.5, also applicable if you lack information. (0 <= p_exp <= 1)}

\item{parameter}{Type TRUE if you do know the populations sd, type FALSE (default) if it is an estimate.}

\item{N}{A positive integer indicating the number of elements in the population. By default, infinite.}
}
\value{
The function returns the sample size needed to estimate the proportion of occurrence of a phenomena, consistent with the risk ('C' and 'e') that the auditor is willing to assume, when conducting a Simple Random Sampling plan, among others with some restrictions.
}
\description{
Sample size needed to estimate the proportion with a Simple Random Sampling plan
}
\examples{
np_srs(C = 0.90, e = 0.04)
np_srs(C = 0.90, e = 0.04, p_exp = 0.5)
np_srs(C = 0.90, e = 0.04, p_exp = 0.5, parameter = TRUE, N = 20000)
np_srs(C = 0.90, e = 0.04, N = 20000)
}
