% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_srs.R
\name{np_srs}
\alias{np_srs}
\title{Sample size n for simple random sampling of proportion}
\usage{
np_srs(C, e, p = 0.5, N = Inf)
}
\arguments{
\item{C}{Level of confidence (0 <= C <= 1)}

\item{e}{Sampling error (0 <= e <= 1).}

\item{p}{Expected proportion in the population. By default 0.5, also applicable if you lack information. (0 <= p <= 1)}

\item{N}{Population size, by default is infinite. Must be a positive integer.}
}
\value{
The function returns the sample size needed to estimate the proportion of occurrence of a phenomena, consistent with the risk ('C' and 'e') that the auditor is willing to assume.
}
\description{
Sample size n for simple random sampling of proportion
}
\examples{
np_srs(0.95, 0.03)
np_srs(0.95, 0.03, 0.5)
np_srs(0.95, 0.03, 0.5, 1000)
np_srs(0.95, 0.03, N = 1000)
}
