% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nx_cls.R
\name{nx_cls}
\alias{nx_cls}
\title{Sample size required to estimate the mean using a cluster sampling design}
\usage{
nx_cls(C, E, sd_exp, m, N = Inf, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{E}{Sampling error; E > 0.}

\item{sd_exp}{Expected standard deviation; sd_exp > 0.}

\item{m}{Average cluster size (number of elements), either of the population -preferred- or of a preliminary sample; m > 0.}

\item{N}{A positive integer indicating the population size (id est, the total number of clusters). Defaults to infinite.}

\item{parameter}{Type TRUE if you do know the populations sd, type FALSE (default) if it is an estimate.}
}
\value{
This function returns the sample size required to estimate the mean of a variable when using a cluster sampling design without replacement, given the level of risk.
}
\description{
Sample size required to estimate the mean using a cluster sampling design
}
\examples{
nx_cls(C = 0.95, E = 650, sd_exp = 15000, m = 8, N = 500)
}
