% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nx_srs.R
\name{nx_srs}
\alias{nx_srs}
\title{Sample size required to estimate the mean using a simple random sampling design}
\usage{
nx_srs(C, E, sd_exp, N = Inf, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{E}{Sampling error; E > 0.}

\item{sd_exp}{Expected standard deviation; sd_exp > 0.}

\item{N}{A positive integer representing the number of elements in the population. Defaults to infinite.}

\item{parameter}{Type TRUE if you do know the population SD, or type FALSE (default) if it is an estimate.}
}
\value{
This function returns the sample size required to estimate the mean of a variable when using a simple random sampling design without replacement, given the level of risk.
}
\description{
Sample size required to estimate the mean using a simple random sampling design
}
\examples{
nx_srs(C = 0.95, E = 50, sd_exp = 400)
nx_srs(C = 0.95, E = 50, sd_exp = 400, parameter = TRUE, N = 10000)
nx_srs(C = 0.95, E = 50, sd_exp = 400, N = 10000, parameter = FALSE)
nx_srs(C = 0.95, E = 100, sd_exp = 400, N = 10000)
}
