% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nx_sts.R
\name{nx_sts}
\alias{nx_sts}
\title{Sample size required to estimate the mean using a stratified sampling design}
\usage{
nx_sts(C, E, sd_exp, alloc = NULL, N, parameter = FALSE)
}
\arguments{
\item{C}{Level of confidence; 0 <= C <= 1.}

\item{E}{Sampling error; E > 0.}

\item{sd_exp}{A vector with the expected standard deviation in each stratum; sd_exp(i) > 0, for every 'i' stratum.}

\item{alloc}{A vector with the relative allocation of sample size for each stratum; 0 < alloc(i) < 1, where sum(alloc(i)) = 1. If not defined (default), its values would be proportional to the size of each stratum.}

\item{N}{A vector of positive integers representing the number of elements in each stratum.}

\item{parameter}{Type TRUE if you do know the populations SD in sd_exp, or type FALSE (default) if they are estimates.}
}
\value{
This function returns the sample size required to estimate the mean of a variable when using a stratified sampling design without replacement, given the level of risk.
}
\description{
Sample size required to estimate the mean using a stratified sampling design
}
\examples{
nx_sts(C = 0.95, E = 2, sd_exp = c(5, 15), N = c(220, 350), parameter = TRUE)
nx_sts(C = 0.95, E = 2, sd_exp = c(5, 15, 10), alloc = c(0.5, 0.2, 0.3), N = c(150, 40, 110), parameter = FALSE)
}
